# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BSYSTEM Integration Core API
  version: 0.8.0
  summary: Authorization-aware normalized API for the BSYSTEM Platform.
  description: |
    The Integration Core exposes a normalized view of the systems BSYSTEM
    integrates with. Source systems stay authoritative; this API never
    re-exposes a raw upstream payload.

    Three surfaces are served:

    - `/api/v1/*` — the human API, authenticated with a user access token.
    - `/api/service/v1/*` — the machine API, authenticated with a service
      identity in the `BSYSTEM-Services` group.
    - `/health`, `/readyz`, `/metrics` — operational endpoints, unauthenticated
      by design, which disclose no business data.

    Authorization is enforced in the backend for every request and denies by
    default. An unknown or unmapped resource results in denial, never in
    broader access.

    Every response carries `X-Request-ID`, echoing the caller's value or a
    generated one, and the same identifier reaches the audit trail and
    published events.

    This document is contract-tested against the server's own route inventory:
    an endpoint cannot be added or removed without the documentation changing
    with it.
  contact:
    name: BSYSTEM Platform
    url: https://github.com/ekucher/bsystem-integration-core
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: The Integration Core, behind the platform reverse proxy.
security:
  - bearerAuth: []
...
paths:
...
  /api/service/v1/whoami:
    get:
      tags: [Service]
      operationId: serviceWhoAmI
      summary: Return the authenticated service identity
      description: |
        Allocates the caller's `SVC-*` Global ID on first sight and returns it
        unchanged afterwards. A human token is refused here, and a token
        outside `BSYSTEM-Services` is refused too.
      security:
        - serviceAuth: []
      responses:
        '200':
          description: The service principal and its resolved authorization.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ServicePrincipal'}
              example:
                id: SVC-000001
                subject: ak-service-3b7d
                name: BSYSTEM Core Service
                username: svc-core
                groups: [BSYSTEM-Services]
                roles: [Service Core]
                permissions: [adapters.read, events.publish, global_ids.read]
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/ServiceForbidden'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 access token
      description: A human access token issued by authentik.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 access token
      description: |
        A service identity access token issued by authentik. The identity must
        be a member of `BSYSTEM-Services`.
...
  responses:
...
    Unauthorized:
      description: The access token is missing, invalid or expired.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: invalid or expired token}
...
    ServiceForbidden:
      description: |
        The token is not a service identity in `BSYSTEM-Services`, or the
        identity lacks the required permission.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {error: service identity group required}
...
    DependencyUnavailable:
      description: |
        A platform dependency is unavailable, or an integration this endpoint
        needs is not configured.

        On an adapter-backed endpoint this is the answer a deployment gets when
        it deliberately leaves a source system out, which is a supported
        configuration rather than a fault. Those answers carry
        `adapter_not_configured` or `capability_unsupported` and name the
        adapter in `source`, so a caller can say which integration is missing
        instead of reporting an unexplained failure. A dependency of the
        platform itself, such as the database, carries no code.

        This is never rendered as an empty collection. An empty result would
        tell a caller the platform knows of no records, when the truth is that
        nobody has told it where to look.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            adapterNotConfigured:
              value: {error: EspoCRM adapter not configured, code: adapter_not_configured, source: espocrm}
            platformDependency:
              value: {error: RBAC store unavailable}
  schemas:
...
    Error:
      type: object
      description: |
        The single error shape for every failure. `error` is always present
        and human-readable; `code` is the stable machine-readable form where
        one is defined. No field ever carries a credential, an internal
        hostname, a stack trace or an upstream payload.
      required: [error]
      properties:
        error: {type: string, description: Human-readable summary.}
        code:
          type: string
          description: |
            Stable machine-readable code. A caller should branch on this
            rather than on the human-readable summary, which may change.

            `upstream_unavailable` — a source system failed, timed out or
            rate-limited the platform.
            `adapter_not_configured` — the integration this endpoint reads
            from is not configured in this deployment. `source` names it. This
            is a deliberate configuration, not a fault.
            `capability_unsupported` — the integration is configured but its
            adapter does not implement this capability. `source` names it.
            `permission_required` — the caller does not hold the permission.
            `scope_required` — the caller holds it but may only exercise it
            inside a scope they have been granted, and no grant matches.
            `not_found` — the resource does not exist, or the caller may not
            learn that it does.
            `invalid_cursor` — the pagination cursor was not issued by this
            platform.
            `invalid_query` — the search text is too short to run.
            `invalid_type` — an unknown entity type was requested.
            `invalid_document` — a search document cannot be indexed as given.
            `search_unavailable` — the search provider could not be reached.
            This is never rendered as an empty result.
            `invalid_event` — an operations event outside the platform's
            vocabulary.
            `invalid_environment` — an unrecognised deployment environment.
            `invalid_relation` — a Global ID that does not resolve, or does
            not name an entity of the expected type.
            `invalid_kind`, `invalid_severity`, `invalid_status` — a support
            value outside its vocabulary.
            `invalid_transition` — a status change the support lifecycle does
            not allow.
            `invalid_source` — an AI source names an entity type the gateway
            does not accept.
            `source_forbidden` — the caller may not use a named source as
            context, or it does not resolve. The two are the same answer.
            `credential_context` — the requested context is
            credential-classified and cannot be sent to a model.
            `prompt_too_large`, `too_many_sources` — the request exceeds the
            gateway's bounds.
            `ai_unavailable` — the model provider could not be reached.
            `unsupported_entity_type` — a Global ID was requested for an
            entity type the platform mints no prefix for. The one allocation
            failure the caller caused and can correct.
            `invalid_principal_type` — a scope grant names a principal kind
            outside `user`, `service`, `group`. Two vocabularies live in this
            schema and are easy to confuse: a role's *kind* is `human` or
            `service`, a scope grant's *principal* is one of these three.
            `invalid_scope_type` — a scope outside `global`, `tenant`,
            `client`, `project`, `resource`.
            `unknown_permission` — a scope grant names a permission the
            catalogue does not hold.
            `scope_store_unavailable` — the grant could not be stored. Like
            the allocation failure below, the summary carries no database
            detail and a `request_id` instead.
            `global_id_allocation_failed` — the platform could not allocate.
            The summary is deliberately free of database detail and carries a
            `request_id` instead: the cause is in the platform's own logs,
            correlated by that id, rather than in the answer.
          enum:
            - upstream_unavailable
            - unsupported_entity_type
            - global_id_allocation_failed
            - invalid_principal_type
            - invalid_scope_type
            - unknown_permission
            - scope_store_unavailable
            - adapter_not_configured
            - capability_unsupported
            - scope_required
            - permission_required
            - not_found
            - invalid_cursor
            - invalid_query
            - invalid_type
            - invalid_document
            - search_unavailable
            - invalid_event
            - invalid_environment
            - invalid_relation
            - invalid_kind
            - invalid_severity
            - invalid_status
            - invalid_transition
            - invalid_source
            - source_forbidden
            - credential_context
            - prompt_too_large
            - too_many_sources
            - ai_unavailable
        source:
          type: string
          description: The adapter id of the failing upstream, when applicable.
        request_id:
          type: string
          description: Correlation id, matching the `X-Request-ID` response header.
      examples:
        - {error: upstream service unavailable, code: upstream_unavailable, source: redmine}
...
    ServicePrincipal:
      type: object
      required: [id, subject, groups, roles, permissions]
      properties:
        id: {type: string, description: Immutable `SVC-*` Global ID., pattern: '^SVC-[0-9]+$'}
        subject: {type: string}
        name: {type: string}
        username: {type: string}
        groups: {type: array, items: {type: string}}
        roles: {type: array, items: {type: string}}
        permissions: {type: array, items: {type: string}}
...